requirements.txt
fastapi
openai
openai-whisper
opencv-python-headless
pydantic
pydantic-settings
python-multipart
torch
uvicorn[standard]
